# tests/OpenCsms.Suite/opencsms.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenCsms.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/tariffs": {
      "post": {
        "tags": [
          "Csms"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterTariffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TariffResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Csms"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TariffResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/stations": {
      "post": {
        "tags": [
          "Csms"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterStationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/stations/{id}": {
      "get": {
        "tags": [
          "Csms"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/stations/{id}/sessions": {
      "get": {
        "tags": [
          "Csms"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/stations/{id}/remote-start": {
      "post": {
        "tags": [
          "Csms"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoteStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoteCommandResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}/remote-stop": {
      "post": {
        "tags": [
          "Csms"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoteCommandResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/invoices/{id}": {
      "get": {
        "tags": [
          "Csms"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/dashboard/stations/{id}": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardStationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/stations/{id}/sessions": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/dashboard/invoices/{id}": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/dashboard/stations/{id}/remote-start": {
      "post": {
        "tags": [
          "Dashboard"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoteStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoteCommandResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/sessions/{id}/remote-stop": {
      "post": {
        "tags": [
          "Dashboard"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoteCommandResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/tariffs/{id}": {
      "put": {
        "tags": [
          "Dashboard"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTariffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TariffResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DashboardStationResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "chargePointId": {
            "type": "string",
            "nullable": true
          },
          "connectorCount": {
            "type": "integer",
            "format": "int32"
          },
          "tariffId": {
            "type": "string",
            "format": "uuid"
          },
          "lastSeenAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "InvoiceResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string",
            "nullable": true
          },
          "energyKwh": {
            "type": "number",
            "format": "double"
          },
          "energyAmount": {
            "type": "number",
            "format": "double"
          },
          "startFeeAmount": {
            "type": "number",
            "format": "double"
          },
          "idleHours": {
            "type": "integer",
            "format": "int32"
          },
          "idleFeeAmount": {
            "type": "number",
            "format": "double"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "issuedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "RegisterStationRequest": {
        "type": "object",
        "properties": {
          "chargePointId": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "connectorCount": {
            "type": "integer",
            "format": "int32"
          },
          "tariffId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "RegisterTariffRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "energyPricePerKwh": {
            "type": "number",
            "format": "double"
          },
          "startFee": {
            "type": "number",
            "format": "double"
          },
          "idleFeePerHour": {
            "type": "number",
            "format": "double"
          },
          "idleGracePeriod": {
            "type": "string",
            "format": "date-span",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RemoteCommandResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RemoteStartRequest": {
        "type": "object",
        "properties": {
          "idTag": {
            "type": "string",
            "nullable": true
          },
          "connectorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SessionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "transactionId": {
            "type": "integer",
            "format": "int32"
          },
          "tenantId": {
            "type": "string",
            "nullable": true
          },
          "stationId": {
            "type": "string",
            "format": "uuid"
          },
          "connectorId": {
            "type": "integer",
            "format": "int32"
          },
          "startedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "endedAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "energyKwh": {
            "type": "number",
            "format": "double"
          },
          "isOpen": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "StationResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string",
            "nullable": true
          },
          "chargePointId": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "connectorCount": {
            "type": "integer",
            "format": "int32"
          },
          "tariffId": {
            "type": "string",
            "format": "uuid"
          },
          "lastSeenAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TariffResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "energyPricePerKwh": {
            "type": "number",
            "format": "double"
          },
          "startFee": {
            "type": "number",
            "format": "double"
          },
          "idleFeePerHour": {
            "type": "number",
            "format": "double"
          },
          "idleGracePeriod": {
            "type": "string",
            "format": "date-span"
          },
          "currency": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateTariffRequest": {
        "type": "object",
        "properties": {
          "energyPricePerKwh": {
            "type": "number",
            "format": "double"
          },
          "startFee": {
            "type": "number",
            "format": "double"
          },
          "idleFeePerHour": {
            "type": "number",
            "format": "double"
          },
          "idleGracePeriod": {
            "type": "string",
            "format": "date-span"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}